move block fee verification to customheader
