+++ plugin/evm/customheader/block_gas_cost_test.go
@@ -7,6 +7,8 @@ import (
 	"math/big"
 	"testing"
 
+	"github.com/ava-labs/libevm/common"
+	"github.com/ava-labs/libevm/common/math"
 	"github.com/ava-labs/libevm/core/types"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
@@ -291,6 +293,171 @@ func TestEstimateRequiredTip(t *testing.T) {
 			requiredTip, err := EstimateRequiredTip(config, test.header)
 			require.ErrorIs(err, test.wantErr)
 			require.Equal(test.want, requiredTip)
+		})
+	}
+}
+
+func TestVerifyBlockFee(t *testing.T) {
+	tests := map[string]struct {
+		baseFee                *big.Int
+		parentBlockGasCost     *big.Int
+		timeElapsed            uint64
+		txs                    []*types.Transaction
+		receipts               []*types.Receipt
+		extraStateContribution *big.Int
+		shouldErr              bool
+	}{
+		"tx only base fee": {
+			baseFee:            big.NewInt(100),
+			parentBlockGasCost: big.NewInt(0),
+			timeElapsed:        0,
+			txs: []*types.Transaction{
+				types.NewTransaction(0, common.HexToAddress("7ef5a6135f1fd6a02593eedc869c6d41d934aef8"), big.NewInt(0), 100, big.NewInt(100), nil),
+			},
+			receipts: []*types.Receipt{
+				{GasUsed: 1000},
+			},
+			extraStateContribution: nil,
+			shouldErr:              true,
+		},
+		"tx covers exactly block fee": {
+			baseFee:            big.NewInt(100),
+			parentBlockGasCost: big.NewInt(0),
+			timeElapsed:        0,
+			txs: []*types.Transaction{
+				types.NewTransaction(0, common.HexToAddress("7ef5a6135f1fd6a02593eedc869c6d41d934aef8"), big.NewInt(0), 100_000, big.NewInt(200), nil),
+			},
+			receipts: []*types.Receipt{
+				{GasUsed: 100_000},
+			},
+			extraStateContribution: nil,
+			shouldErr:              false,
+		},
+		"txs share block fee": {
+			baseFee:            big.NewInt(100),
+			parentBlockGasCost: big.NewInt(0),
+			timeElapsed:        0,
+			txs: []*types.Transaction{
+				types.NewTransaction(0, common.HexToAddress("7ef5a6135f1fd6a02593eedc869c6d41d934aef8"), big.NewInt(0), 100_000, big.NewInt(200), nil),
+				types.NewTransaction(1, common.HexToAddress("7ef5a6135f1fd6a02593eedc869c6d41d934aef8"), big.NewInt(0), 100_000, big.NewInt(100), nil),
+			},
+			receipts: []*types.Receipt{
+				{GasUsed: 100_000},
+				{GasUsed: 100_000},
+			},
+			extraStateContribution: nil,
+			shouldErr:              false,
+		},
+		"txs split block fee": {
+			baseFee:            big.NewInt(100),
+			parentBlockGasCost: big.NewInt(0),
+			timeElapsed:        0,
+			txs: []*types.Transaction{
+				types.NewTransaction(0, common.HexToAddress("7ef5a6135f1fd6a02593eedc869c6d41d934aef8"), big.NewInt(0), 100_000, big.NewInt(150), nil),
+				types.NewTransaction(1, common.HexToAddress("7ef5a6135f1fd6a02593eedc869c6d41d934aef8"), big.NewInt(0), 100_000, big.NewInt(150), nil),
+			},
+			receipts: []*types.Receipt{
+				{GasUsed: 100_000},
+				{GasUsed: 100_000},
+			},
+			extraStateContribution: nil,
+			shouldErr:              false,
+		},
+		"split block fee with extra state contribution": {
+			baseFee:            big.NewInt(100),
+			parentBlockGasCost: big.NewInt(0),
+			timeElapsed:        0,
+			txs: []*types.Transaction{
+				types.NewTransaction(0, common.HexToAddress("7ef5a6135f1fd6a02593eedc869c6d41d934aef8"), big.NewInt(0), 100_000, big.NewInt(150), nil),
+			},
+			receipts: []*types.Receipt{
+				{GasUsed: 100_000},
+			},
+			extraStateContribution: big.NewInt(5_000_000),
+			shouldErr:              false,
+		},
+		"extra state contribution insufficient": {
+			baseFee:                big.NewInt(100),
+			parentBlockGasCost:     big.NewInt(0),
+			timeElapsed:            0,
+			txs:                    nil,
+			receipts:               nil,
+			extraStateContribution: big.NewInt(9_999_999),
+			shouldErr:              true,
+		},
+		"negative extra state contribution": {
+			baseFee:                big.NewInt(100),
+			parentBlockGasCost:     big.NewInt(0),
+			timeElapsed:            0,
+			txs:                    nil,
+			receipts:               nil,
+			extraStateContribution: big.NewInt(-1),
+			shouldErr:              true,
+		},
+		"extra state contribution covers block fee": {
+			baseFee:                big.NewInt(100),
+			parentBlockGasCost:     big.NewInt(0),
+			timeElapsed:            0,
+			txs:                    nil,
+			receipts:               nil,
+			extraStateContribution: big.NewInt(10_000_000),
+			shouldErr:              false,
+		},
+		"extra state contribution covers more than block fee": {
+			baseFee:                big.NewInt(100),
+			parentBlockGasCost:     big.NewInt(0),
+			timeElapsed:            0,
+			txs:                    nil,
+			receipts:               nil,
+			extraStateContribution: big.NewInt(10_000_001),
+			shouldErr:              false,
+		},
+		"tx only base fee after full time window": {
+			baseFee:            big.NewInt(100),
+			parentBlockGasCost: big.NewInt(500_000),
+			timeElapsed:        ap4.TargetBlockRate + 10,
+			txs: []*types.Transaction{
+				types.NewTransaction(0, common.HexToAddress("7ef5a6135f1fd6a02593eedc869c6d41d934aef8"), big.NewInt(0), 100, big.NewInt(100), nil),
+			},
+			receipts: []*types.Receipt{
+				{GasUsed: 1000},
+			},
+			extraStateContribution: nil,
+			shouldErr:              false,
+		},
+		"tx only base fee after large time window": {
+			baseFee:            big.NewInt(100),
+			parentBlockGasCost: big.NewInt(100_000),
+			timeElapsed:        math.MaxUint64,
+			txs: []*types.Transaction{
+				types.NewTransaction(0, common.HexToAddress("7ef5a6135f1fd6a02593eedc869c6d41d934aef8"), big.NewInt(0), 100, big.NewInt(100), nil),
+			},
+			receipts: []*types.Receipt{
+				{GasUsed: 1000},
+			},
+			extraStateContribution: nil,
+			shouldErr:              false,
+		},
+	}
+
+	for name, test := range tests {
+		t.Run(name, func(t *testing.T) {
+			blockGasCost := BlockGasCostWithStep(
+				test.parentBlockGasCost,
+				ap4.BlockGasCostStep,
+				test.timeElapsed,
+			)
+			bigBlockGasCost := new(big.Int).SetUint64(blockGasCost)
+
+			if err := VerifyBlockFee(test.baseFee, bigBlockGasCost, test.txs, test.receipts, test.extraStateContribution); err != nil {
+				if !test.shouldErr {
+					t.Fatalf("Unexpected error: %s", err)
+				}
+			} else {
+				if test.shouldErr {
+					t.Fatal("Should have failed verification")
+				}
+			}
 		})
 	}
 }
